requirements.txt
requests
flask
nanoid
pygments
websocket-client
